+++ config/generator/icons/generateIcons.js
@@ -11,8 +11,8 @@ const execAsync = util.promisify(exec);
 
 const applyPrettier = async () =>
   Promise.all([
-    execAsync('npx pretty-quick --pattern "src/atoms/icons/*.tsx"'),
-    execAsync('npx pretty-quick --pattern "src/atoms/icons/index.ts"')
+    execAsync('npx prettier --write "src/atoms/icons/*.tsx"'),
+    execAsync('npx prettier --write "src/atoms/icons/index.ts"')
   ]);
 
 const generateIcon = async (name, path) => {