+++ src/atoms/tooltip/Tooltip.tsx
@@ -3,18 +3,18 @@ import styled from 'styled-components';
 import { white } from '@colors';
 import { Placement } from '@atoms/tooltip/types';
 import {
-  checkIsStyleComputed,
-  computeTooltipHeight,
-  computeTooltipWidth,
-  getAlternativeStyle,
-  getBoundingRect,
-  getCSSComputedStyle,
+  displayIfInContainer,
   getTooltipBgColorByType,
   handleTooltipArrowPosition,
   handleTooltipPosition,
-  isElementOutOfContainer,
-  removeOutOfScreenPlacement
+  isTooltipOutOfContainer
 } from '@atoms/tooltip/helpers';
+import {
+  getBoundingRect,
+  getContainerBoundaries,
+  getElementDimensions,
+  getElementPlacement
+} from '@utility/positionCompute';
 
 export interface Props {
   title: string;
@@ -40,15 +40,18 @@ export interface Props {
     /**
      * An element as a container (if not the default window)
      */
-    container?: HTMLElement | null;
+    container?: React.RefObject<HTMLElement>;
   };
 }
 
 interface State {
-  placement: Placement;
+  outOfContainer: boolean;
+  placement: Placement | null;
 }
 
-export type WrapperProps = Omit<Required<Props>, 'title'>;
+export type WrapperProps = Omit<Required<Props>, 'title'> & {
+  outOfContainer: boolean;
+};
 
 const TooltipWrapper = styled.span<WrapperProps>`
   > :first-child:hover {
@@ -85,8 +88,7 @@ const TooltipWrapper = styled.span<WrapperProps>`
   }
 
   > :first-child:hover::before {
-    visibility: visible;
-    opacity: 1;
+    ${displayIfInContainer}
   }
 
   > :first-child::after {
@@ -107,13 +109,12 @@ const TooltipWrapper = styled.span<WrapperProps>`
   }
 
   > :first-child:hover::after {
-    visibility: visible;
-    opacity: 1;
+    ${displayIfInContainer}
   }
 `;
 
 class Tooltip extends React.PureComponent<Props, State> {
-  private ref: EventTarget | null;
+  private ref: React.RefObject<HTMLElement> = React.createRef();
   private coordinates: { top: number | null; left: number | null };
   private readonly tooltipPossiblePlacements: Placement[] = [];
   private readonly defaultPlacement: Placement;
@@ -121,9 +122,9 @@ class Tooltip extends React.PureComponent<Props, State> {
   public constructor(props: Props) {
     super(props);
     this.defaultPlacement = 'top';
-    this.ref = null;
     this.state = {
-      placement: props.placement || this.defaultPlacement
+      placement: props.placement || this.defaultPlacement,
+      outOfContainer: false
     };
     this.tooltipPossiblePlacements = ['top', 'left', 'right', 'bottom'];
     this.coordinates = {
@@ -140,7 +141,7 @@ class Tooltip extends React.PureComponent<Props, State> {
     this.positionTooltip();
   }
 
-  private onHover = (event: MouseEvent) => {
+  private onHover = (event: React.MouseEvent<HTMLElement>) => {
     const target = event.currentTarget;
     const { top: previousTop, left: previousLeft } = this.coordinates;
     const { top, left } = getBoundingRect(target);
@@ -150,72 +151,49 @@ class Tooltip extends React.PureComponent<Props, State> {
     }
 
     this.coordinates = { ...this.coordinates, ...{ top, left } };
-    this.ref = target;
     this.positionTooltip();
   };
 
-  private getTooltipPlacement(
-    placement: Placement,
-    possibilities: Placement[]
-  ): Placement {
-    if (!placement) {
-      throw new Error(
-        'Tooltip cannot be displayed in the container Element/Window due to its incompatible size. Please consider reducing it or better position tooltiped element in the screen.'
-      );
-    }
-
-    if (!this.isOutScreen(placement)) {
-      return placement;
-    }
-
-    const newPossibilities = removeOutOfScreenPlacement(
-      possibilities,
-      placement
-    );
-    const [firstPossiblePlacement] = newPossibilities;
-
-    return this.getTooltipPlacement(firstPossiblePlacement, newPossibilities);
-  }
-
-  private isOutScreen(placement: Placement): boolean {
-    const tooltipStyle = getCSSComputedStyle(this.ref);
-    const isStyleComputed = checkIsStyleComputed(tooltipStyle);
-    const { width = null, height = null } = isStyleComputed
-      ? {}
-      : getAlternativeStyle(tooltipStyle, this.ref); // Extra calculations For Edge
-
-    const totalWidth = computeTooltipWidth(tooltipStyle, placement, width);
-    const totalHeight = computeTooltipHeight(tooltipStyle, placement, height);
-    const rect = getBoundingRect(this.ref);
-    const { options } = this.props;
-    const container =
-      !options || !options.container ? window : options.container;
-
-    return isElementOutOfContainer({
-      rect,
-      measurements: { totalWidth, totalHeight },
-      container,
-      placement
-    });
-  }
-
   private positionTooltip() {
+    if (!this.ref.current) {
+      return;
+    }
+    const isBeforePseudoElement = true;
+    const elementDimensions = getElementDimensions({
+      elementRef: this.ref,
+      elementIsBefore: isBeforePseudoElement
+    });
+
+    const containerElement =
+      (this.props.options &&
+        this.props.options.container &&
+        this.props.options.container.current) ||
+      window;
+    const containerDimensions = getContainerBoundaries(containerElement);
     try {
-      const placement = this.getTooltipPlacement(
-        this.props.placement || this.defaultPlacement,
-        this.tooltipPossiblePlacements
-      );
+      const placement = getElementPlacement({
+        isElementOutOfContainerMethod: isTooltipOutOfContainer,
+        placement: this.props.placement || this.defaultPlacement,
+        possibilities: this.tooltipPossiblePlacements,
+        elementDimensions,
+        containerDimensions
+      });
       this.setState({ placement: placement });
     } catch (e) {
-      console.error(e);
+      this.setState({ outOfContainer: true });
+      console.error('Error while computing Tooltip position', e);
     }
   }
 
-  private cloneChildren = (title: string) => (child: React.ReactNode) => {
+  private cloneChildren = (title: string) => (
+    child: React.ReactNode,
+    childIndex: number
+  ) => {
     if (React.isValidElement(child)) {
       return React.cloneElement(child, {
         'data-tooltip': title,
-        onMouseEnter: this.onHover
+        onMouseEnter: this.onHover,
+        ref: childIndex === 0 && this.ref
       });
     }
   };
@@ -228,14 +206,15 @@ class Tooltip extends React.PureComponent<Props, State> {
       options = {}
     } = this.props;
     const childrenWithProps = React.Children.map(
-      children,
+      <div>{children}</div>, // Wrapping the children with a div to make sure that the tooltip will be attached to this div so any state modification of the initial children (disabled, etc..) won't affect the style of the tooltip
       this.cloneChildren(title)
     );
 
     return (
       <TooltipWrapper
         type={type}
-        placement={this.state.placement}
+        placement={this.state.placement || this.defaultPlacement}
+        outOfContainer={this.state.outOfContainer}
         options={options}
       >
         {childrenWithProps}