+++ src/molecules/index.ts
@@ -1,0 +1,3 @@
+export {
+  default as QuickActionsMenu
+} from './quick-actions-menu/QuickActionsMenu';